backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
channels==4.3.2
channels_redis==4.3.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt
PyJWT
pytz
psycopg2-binary
python-dotenv
msgpack==1.1.2
redis==7.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
